实现了积分兑换逻辑和后端接口（积分转换），bug修复

+++ credit-bank-mvp/src/main/java/com/creditbank/mvp/service/ExchangeRuleService.java
@@ -99,7 +99,7 @@ public class ExchangeRuleService {
         }
 
         if (rule.getPerUserLimit() != null && rule.getPerUserLimit() > 0) {
-            int exchangedCount = transactionLogMapper.selectCount(
+            Long exchangedCount = transactionLogMapper.selectCount(
                     new LambdaQueryWrapper<TransactionLog>()
                             .eq(TransactionLog::getUserId, userId)
                             .eq(TransactionLog::getBizType, "EXCHANGE")
